Changes
Reorganized the script files. now there is an extra folder with scripts
not in use (didn't have the courage to delete them).

To make the game work with a singular magnetscript I had to make a few
changes to the c# file: changes check and assign the appropriate charge
to their assigned racer (gameobject)